refactor!: Move libvcs.util -> libvcs.cmd.core
