+++ libvcs/conftest.py
@@ -10,8 +10,8 @@ import pytest
 
 from faker import Faker
 
+from libvcs.cmd.core import run, which
 from libvcs.states.git import GitRemoteDict, GitRepo
-from libvcs.util import run, which
 
 skip_if_git_missing = pytest.mark.skipif(
     not which("git"), reason="git is not available"
